исправил LazyColum   и сделал так чтобы моделька character работал с lazyColum
