+++ app/src/main/java/com/example/teamworklesson3compose/presentation/data/remote/models/persons/Result.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Result(
     @SerializedName("age")
-    val age: Int,
+    val age: String,
     @SerializedName("alias")
     val alias: List<String>,
     @SerializedName("birthplace")
